main/requirements.txt
discord-py-interactions==5.2.0
discord-py-slash-command~=4.2.1
python-dateutil~=2.8.2
python-dotenv~=1.0.0
regex~=2023.5.5
typing-extensions~=4.5.0